Add fetch to recipes
