+++ src/scripts.js
@@ -2,8 +2,9 @@ import './css/base.scss';
 import './css/styles.scss';
 
 import domUpdates from './domUpdates';
-import recipeData from './data/recipes';
-import ingredientsData from './data/ingredients';
+// import recipeData from './data/recipes';
+// For ingredientData create a fetch call and assign it to a global variable also for (recipesData?)'
+// import ingredientsData from './data/ingredients';
 // import users from './data/users';
 import { fetchApiCalls } from './apiCalls'
 
@@ -20,44 +21,72 @@ let inputSearch = document.querySelector('.search-input');
 let checkBoxes = document.querySelectorAll("input[type=checkbox]");
 let submitTagsButton = document.querySelector('#submitTagsButton');
 let cardArea = document.querySelector('.all-cards');
-let cookbook = new Cookbook(recipeData);
+// let cookbook = new Cookbook(recipeData);
+/////// create a fetch call to update the recipeData !
+let user, pantry, cookbook;
 
-let user, pantry;
-
-window.onload = generateUser();
-// window.addEventListener('load', generateUser);
-
-// function generateNewUser() {
-//   fetchApiCalls('users').then(data => {
-//     console.log(data)
-//     console.log(user)
-//   })
-// }
 
 homeButton.addEventListener('click', cardButtonConditionals);
 searchButton.addEventListener('click', searchByNameIng);
 submitTagsButton.addEventListener('click', searchByTags);
 favButton.addEventListener('click', viewFavorites);
 toCookButton.addEventListener('click', viewRecipesToCook);
-// This event listener work as a event bubbling
 cardArea.addEventListener('click', cardButtonConditionals);
+window.onload = generateUser();
 
 // Functions
-
 function preventDefault() {
   event.preventDefault()
 }
 
 
+
+
+////////////////////////////////  FETCH CALLS
+
+function generateUser() {
+  fetchApiCalls('users')
+  .then(data => {
+    // console.log('hola', data.usersData.length);
+    const randomUserNum = Math.floor(Math.random() * data.usersData.length);
+    let matchingUser = data.usersData.find((item) => {
+      if (item.id === randomUserNum) {
+        return item;
+      }
+    })
+    user = new User(matchingUser);
+    // console.log(user);
+    pantry = new Pantry(user.pantry)
+    domUpdates.displayGreetUser(user);
+    // populateCards(cookbook.recipes);
+    recipeData()
+  })
+}
+
+
+function recipeData(){
+ let recipePromise = fetchApiCalls('recipes')
+ .then(data => {
+   let recipeData = data.recipeData;
+
+  let cookbook = new Cookbook(data.recipeData);
+  populateCards(cookbook.recipes);
+    // console.log('ing', recipeData);
+ })
+}
+
+
+
+
+//////  FILTER RECIPES BY - TAGS & TEST (NAME & ING)
 function searchByNameIng() {
   preventDefault()
-
   // Works for SEARCH BY ALL RECIPES
   // if (!cardArea.classList.contains('all')) {
   //   const searchText = cookbook.findRecipe(inputSearch.value);
   //   populateCards(searchText);
   // }
-
+  // console.log(cookbook.recipes);
   // Works for SEARCH BY FAVORITES
   if (!cardArea.classList.contains('favorite')) {
     const searchTextFav = user.findFavorites(inputSearch.value);
@@ -65,9 +94,7 @@ function searchByNameIng() {
   }
 }
 
-
-
-/////////////////////////////////////////
+//------------------------------------------------------
 /// Function filter recipes by "TAGS" based on (user.favoriteRecipes)
 function searchByTags() {
   preventDefault()
@@ -77,15 +104,11 @@ function searchByTags() {
       checkBoxMatches.push(checkBox.value)
     }
   })
-
   // Works for FILTER BY ALL RECIPES
   if (!cardArea.classList.contains('all')) {
     let tagMatches = cookbook.filterRecipesTags(checkBoxMatches);
     populateCards(tagMatches);
   }
-
-
-
 // Works for FILTER BY FAVORITES
   // if (!cardArea.classList.contains('all')) {
   //   let tagMatches = user.filterFavorites(checkBoxMatches);
@@ -95,69 +118,17 @@ function searchByTags() {
 }
 
 
-// function onStartup() {
-//   generateUser()
-//   // domUpdates.displayGreetUser(user);
-//     // pantry = new Pantry(user.pantry)
-//     // populateCards(cookbook.recipes);
-//     // domUpdates.displayGreetUser(user);
-//
-// }
 
-// function generateNewUser() {
-//   fetchApiCalls('users').then(data => {
-//     console.log(data)
-//     console.log(user)
-//   })
-// }
 
-function generateUser() {
-  fetchApiCalls('users')
-  .then(data => {
-    console.log('hola', data.usersData.length);
-    const randomUserNum = Math.floor(Math.random() * data.usersData.length);
-    let matchingUser = data.usersData.find((item) => {
-      if (item.id === randomUserNum) {
-        return item;
-      }
-    })
-    user = new User(matchingUser);
-    console.log(user);
-    pantry = new Pantry(user.pantry)
-    populateCards(cookbook.recipes);
-    domUpdates.displayGreetUser(user);
-  })
 
-  // const randomUserNum = Math.floor(Math.random() * data.length);
-  // let matchingUser = data.find((item) => {
-  //   if (item.id === randomUserNum) {
-  //     return item;
-  //   }
-  // })
-  // user = new User(matchingUser);
-  // console.log(user);
-}
 
-// YE OLD CRAPPY CODE THANK YOU
-// function onStartup() {
-//   let userId = (Math.floor(Math.random() * 49) + 1)
-//   let newUser = users.find(user => {
-//     return user.id === Number(userId);
-//   }); // rather than userId === newUser.id
-//   user = new User(userId, newUser.name, newUser.pantry)
-//   pantry = new Pantry(newUser.pantry)
-//   populateCards(cookbook.recipes);
-//   greetUser();
-// }
 
-// FETCH CALLS
-// function generateNewUser() {
-//   fetchApiCalls('users').then(data => {
-//     console.log(data)
-//     console.log('hello')
-//   })
-// }
 
+
+
+
+
+//////////////////////// DOM MANIPULATION ON  ------------ (domUpdates.js)
 function viewFavorites() {
   domUpdates.displayFavorites(user, cardArea, favButton, populateCards, cookbook);
 
@@ -241,43 +212,102 @@ function cardButtonConditionals() {
   // }
 }
 
-function displayDirections(event) {
-    // console.log(user);
-  let newRecipeInfo = cookbook.recipes.find(recipe => {
-    if (recipe.id === Number(event.target.id)) {
-      return recipe;
-    }
-  })
-  let recipeObject = new Recipe(newRecipeInfo, ingredientsData);
-  let cost = recipeObject.calculateCost()
-  let costInDollars = (cost / 100).toFixed(2)
-  let returnInstructions = recipeObject.retrieveRecipeInstructions();
-  let returnIngredients = recipeObject.retrieveIngredientName();
-  console.log(returnIngredients, returnInstructions);
-  // What does it mean "all" ??
-  cardArea.classList.add('all');
-  cardArea.innerHTML = `<h3>${recipeObject.name}</h3>
-  <p class='all-recipe-info'>
-  <strong>It will cost: </strong><span class='cost recipe-info'>
-  $${costInDollars}</span><br><br>
-  <strong>You will need: </strong><span class='ingredients recipe-info'></span>
-  <strong>Instructions: </strong><ol><span class='instructions recipe-info'>
-  </span></ol>
-  </p>`;
-  let ingredientsSpan = document.querySelector('.ingredients');
-  let instructionsSpan = document.querySelector('.instructions');
-  recipeObject.ingredients.forEach(ingredient => {
-    ingredientsSpan.insertAdjacentHTML('afterbegin', `<ul><li>
-    ${ingredient.quantity.amount.toFixed(2)} ${ingredient.quantity.unit}
-    ${ingredient.name}</li></ul>
-    `)
-  })
-  recipeObject.instructions.forEach(instruction => {
-    instructionsSpan.insertAdjacentHTML('beforebegin', `<li>
-    ${instruction.instruction}</li>
-    `)
+
+
+
+
+
+
+
+
+
+////////////////////////// DOM MANIPULATON ON   -------------- (SCRIPT.JS)
+// function displayDirections(event) {
+//     // console.log(user);
+//   let newRecipeInfo = cookbook.recipes.find(recipe => {
+//     if (recipe.id === Number(event.target.id)) {
+//       return recipe;
+//     }
+//   })
+//   let recipeObject = new Recipe(newRecipeInfo, ingredientsData);
+//   let cost = recipeObject.calculateCost()
+//   let costInDollars = (cost / 100).toFixed(2)
+//   let returnInstructions = recipeObject.retrieveRecipeInstructions();
+//   let returnIngredients = recipeObject.retrieveIngredientName();
+//   console.log(returnIngredients, returnInstructions);
+//   /// CALL FUNCTION1 AND FUNCTION2
+//
+//
+//   // MOVE TO DOM UPDATES FILE
+//   cardArea.classList.add('all');
+//   cardArea.innerHTML = `<h3>${recipeObject.name}</h3>
+//   <p class='all-recipe-info'>
+//   <strong>It will cost: </strong><span class='cost recipe-info'>
+//   $${costInDollars}</span><br><br>
+//   <strong>You will need: </strong><span class='ingredients recipe-info'></span>
+//   <strong>Instructions: </strong><ol><span class='instructions recipe-info'>
+//   </span></ol>
+//   </p>`;
+//   ////////////////////////////////////
+//
+//   /// FUNCTION 1
+//
+//
+//   let ingredientsSpan = document.querySelector('.ingredients');
+//   recipeObject.ingredients.forEach(ingredient => {
+//     ingredientsSpan.insertAdjacentHTML('afterbegin', `<ul><li>
+//     ${ingredient.quantity.amount.toFixed(2)} ${ingredient.quantity.unit}
+//     ${ingredient.name}</li></ul>
+//     `)
+//   })
+//
+//
+//   //// FUNCTION  2
+//   let instructionsSpan = document.querySelector('.instructions');
+//   recipeObject.instructions.forEach(instruction => {
+//     instructionsSpan.insertAdjacentHTML('beforebegin', `<li>
+//     ${instruction.instruction}</li>
+//     `)
+//   })
+// }
+
+
+
+
+
+
+/////// NEW FUNCTION WITH FETCH  ALL
+function displayDirections() {
+  fetchApiCalls('ingredients')
+  .then(data => {
+    console.log('test', data);
+
+    let newRecipeInfo = cookbook.recipes.find(recipe => {
+      if (recipe.id === Number(event.target.id)) {
+        return recipe;
+      }
+    })
+
+    let recipeObject = new Recipe(newRecipeInfo, data.ingredientsData);
+    let cost = recipeObject.calculateCost()
+    let costInDollars = (cost / 100).toFixed(2)
+    let returnInstructions = recipeObject.retrieveRecipeInstructions();
+    let returnIngredients = recipeObject.retrieveIngredientName();
+
+    domUpdates.displayCardDirections(cardArea, recipeObject, costInDollars, returnInstructions, returnIngredients)
   })
 }
+
+
+
+
+
+
+
+
+
+
+
 
 
 // This function helps us when we return to the main view and the cards are stil active in the main vuew, need to make the same for the recipes to cook section
